+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Fix: Add tests for all utils functions (@lwasser, #122)
 - Fix: Bug where date_accepted is removed (@lwasser, #129)
 - Fix: Refactor all GitHub related methods move to gh_client module (@lwasser, #125)
+- Add: support for partners and emeritus_editor in contributor model (@lwasser, #133)
 
 
 ## [v0.2.3] - 2024-02-29